allow run, copy, remove edit
